feat: enable seamless linking to bl-web for /info
